reorganize shared/ again
