+++ client/browser/src/libs/code_intelligence/extensions.tsx
@@ -4,8 +4,8 @@ import { combineLatest, from, Observable, Unsubscribable } from 'rxjs'
 import { map, take } from 'rxjs/operators'
 import { ContributableMenu } from '../../../../../shared/src/api/protocol'
 import { TextDocumentDecoration } from '../../../../../shared/src/api/protocol/plainTypes'
-import { Controller as ClientController, createController } from '../../../../../shared/src/client/controller'
 import { CommandListPopoverButton } from '../../../../../shared/src/commandPalette/CommandList'
+import { Controller as ClientController, createController } from '../../../../../shared/src/extensions/controller'
 import { Notifications } from '../../../../../shared/src/notifications/Notifications'
 import { ConfiguredSubject, SettingsCascade, SettingsCascadeOrError } from '../../../../../shared/src/settings'
 
@@ -17,7 +17,7 @@ import {
     decorationStyleForTheme,
 } from '../../../../../shared/src/api/client/providers/decoration'
 import { Context } from '../../../../../shared/src/context'
-import { viewerConfiguredExtensions } from '../../../../../shared/src/controller'
+import { viewerConfiguredExtensions } from '../../../../../shared/src/extensions/helpers'
 import { isErrorLike } from '../../shared/backend/errors'
 import { createExtensionsContext, createMessageTransports } from '../../shared/backend/extensions'
 import { GlobalDebug } from '../../shared/components/GlobalDebug'
