+++ client/browser/src/shared/components/CodeViewToolbar.tsx
@@ -3,8 +3,8 @@ import * as React from 'react'
 import { Subscription } from 'rxjs'
 import { ActionsNavItems } from '../../../../../shared/src/actions/ActionsNavItems'
 import { ContributableMenu } from '../../../../../shared/src/api/protocol'
-import { ControllerProps } from '../../../../../shared/src/client/controller'
 import { ExtensionsContextProps } from '../../../../../shared/src/context'
+import { ControllerProps } from '../../../../../shared/src/extensions/controller'
 import { ISite, IUser } from '../../../../../shared/src/graphqlschema'
 import { SettingsCascadeProps } from '../../../../../shared/src/settings'
 import { FileInfo } from '../../libs/code_intelligence'
