+++ client/browser/src/shared/components/LegacyCodeViewToolbar.tsx
@@ -12,8 +12,8 @@ import * as React from 'react'
 import { Subscription } from 'rxjs'
 import { ActionsNavItems } from '../../../../../shared/src/actions/ActionsNavItems'
 import { ContributableMenu } from '../../../../../shared/src/api/protocol'
-import { ControllerProps } from '../../../../../shared/src/client/controller'
 import { ExtensionsContextProps } from '../../../../../shared/src/context'
+import { ControllerProps } from '../../../../../shared/src/extensions/controller'
 import { ISite, IUser } from '../../../../../shared/src/graphqlschema'
 import { SettingsCascadeProps } from '../../../../../shared/src/settings'
 import { SimpleProviderFns } from '../backend/lsp'
