+++ shared/src/extensions/ExtensionStatus.tsx
@@ -5,9 +5,9 @@ import { Subject, Subscription } from 'rxjs'
 import { distinctUntilChanged, map, switchMap } from 'rxjs/operators'
 import { ExtensionConnection } from '../api/client/controller'
 import { BrowserConsoleTracer, Trace } from '../api/protocol/jsonrpc2/trace'
-import { ControllerProps } from '../client/controller'
 import { PopoverButton } from '../components/PopoverButton'
 import { Toggle } from '../components/Toggle'
+import { ControllerProps } from '../extensions/controller'
 
 interface Props extends ControllerProps {
     link: React.ComponentType<{ id: string }>
