+++ shared/src/extensions/helpers.ts
@@ -1,11 +1,11 @@
 import { from, Observable, of, throwError } from 'rxjs'
 import { catchError, filter, map, startWith, switchMap } from 'rxjs/operators'
-import { Context } from './context'
-import { asError, createAggregateError, ErrorLike, isErrorLike } from './errors'
-import { ConfiguredExtension, toConfiguredExtension } from './extensions/extension'
-import { gql, graphQLContent } from './graphql'
-import * as GQL from './graphqlschema'
-import { SettingsCascadeOrError } from './settings'
+import { Context } from '../context'
+import { asError, createAggregateError, ErrorLike, isErrorLike } from '../errors'
+import { gql, graphQLContent } from '../graphql'
+import * as GQL from '../graphqlschema'
+import { SettingsCascadeOrError } from '../settings'
+import { ConfiguredExtension, toConfiguredExtension } from './extension'
 
 const LOADING: 'loading' = 'loading'
