+++ web/src/SourcegraphWebApp.tsx
@@ -9,10 +9,10 @@ import { startWith } from 'rxjs/operators'
 import { EMPTY_ENVIRONMENT as EXTENSIONS_EMPTY_ENVIRONMENT } from '../../shared/src/api/client/environment'
 import { TextDocumentItem } from '../../shared/src/api/client/types/textDocument'
 import { WorkspaceRoot } from '../../shared/src/api/protocol/plainTypes'
-import { createController as createExtensionsController } from '../../shared/src/client/controller'
-import { viewerConfiguredExtensions } from '../../shared/src/controller'
 import { isErrorLike } from '../../shared/src/errors'
+import { createController as createExtensionsController } from '../../shared/src/extensions/controller'
 import { ConfiguredExtension } from '../../shared/src/extensions/extension'
+import { viewerConfiguredExtensions } from '../../shared/src/extensions/helpers'
 import * as GQL from '../../shared/src/graphqlschema'
 import { Notifications } from '../../shared/src/notifications/Notifications'
 import { ConfiguredSubject, SettingsCascadeOrError } from '../../shared/src/settings'
