+++ web/src/settings/SettingsArea.tsx
@@ -6,8 +6,8 @@ import { Route, RouteComponentProps, Switch } from 'react-router'
 import { combineLatest, Observable, Subject, Subscription } from 'rxjs'
 import { catchError, distinctUntilChanged, map, startWith, switchMap } from 'rxjs/operators'
 import settingsSchemaJSON from '../../../schema/settings.schema.json'
-import { queryConfiguredExtensions } from '../../../shared/src/controller'
 import { createAggregateError, ErrorLike, isErrorLike } from '../../../shared/src/errors'
+import { queryConfiguredExtensions } from '../../../shared/src/extensions/helpers'
 import { gql } from '../../../shared/src/graphql'
 import { ISettingsCascade } from '../../../shared/src/graphqlschema'
 import * as GQL from '../../../shared/src/graphqlschema'
